depmodwrapper-cross: Mark as machine specific
The kernel and its staging directory are machine specific so the wrapper
needs to be as well. Also take the opportunity to remove the default
dependencies of the recipe since its a script and doesn't need the cross
compiler. This gives the build a little more scheduing freedom.

(From OE-Core rev: 1cc2ac769baae19b54184df5f43a3e77c1743c29)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>