bitbake: bb.fetch.git: truncate branches to SRCREV
This removes the commits after SRCREV up to the branch HEAD, as they aren't
needed for us, and just increase the size of the tarball. The user can always
`git fetch` to pull down those commits. Repack/prune the repository before
archiving it to ensure those commits are physically removed.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>