bitbake: fetch2/git: rework shallow repo construction
With this new method, it works with linux-yocto kernel repositories. Rather
than constructing a minimal repo with only the branches specified in the url,
with each made shallow at a specific depth or revision, we assemble a list of
shallow revisions, and process them, also processing any points at which their
history intersects the branches in the repository. This ensures that, for
example, if v4.1.7 is made shallow, nothing prior to it gets pulled in via
other avenues, e.g. v4.1.6. Instead v4.1.7 is made shallow and all of its
history is removed or made shallow, guaranteed. I'll document the method in
more detail when we submit upstream.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>