bitbake: fetch2: compare localpaths, not methods
We only want to trigger the mirror tarball handling if the basenames don't
match up. If we do so based solely on method, then valid cases where a file is
fetched from a different method than the base url will trigger that logic.

In the case where an http file is found via a file mirror, this was causing
the http fetcher to be run to process the nonexistent mirror tarball,
resulting in an attempt to download from upstream when we already have the
file.

JIRA: SB-6156

Signed-off-by: Christopher Larson <chris_larson@mentor.com>