wic: don't throw away our created swap partition
We were creating the partition, mkswap'ing it, and then not dd'ing it into
place in the final image, as source_file wasn't being set for swap partitions.
This would result in a swap partition that couldn't be enabled on target until
mkswap was run.

Upstream: git://git.openembedded.org/openembedded-core
Commit-ID: 7580942a66f84c2120763c2d9665e2cec8e25138

Signed-off-by: Christopher Larson <chris_larson@mentor.com>
Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Srikanth Krishnakar <Srikanth_Krishnakar@mentor.com>