toaster: Use custom script if available instead of oe-init
This is a change to include users to have custom script
instead of oe-init from poky. Since we have local.conf.append,
the oe-init would be a failure. So we take this approach.

In the upstream there is a patch which is under review:
http://git.yoctoproject.org/cgit/cgit.cgi/poky-contrib/log/?h=sujith/toaster-use-custom-build-scriptsV2

Once this patch gets pulled into the upstream branches, we can
safely revert this change.

Signed-off-by: Sujith Haridasan <Sujith_Haridasan@mentor.com>