sysroot-relativelinks: also consider links to dirs on the host
Dead symlinks, or symlinks to existing files will show up in 'files' of an
os.walk, but symlinks to existing directories show up in 'dirs', so we need to
consider both.

As one specific example, the symlink from /usr/lib/ssl/certs was left pointing
to /etc/ssl/certs rather than the relative path when the sdk was built on
hosts where the latter exists.

JIRA: SB-8374

Signed-off-by: Christopher Larson <chris_larson@mentor.com>